+++ todolist/todolist.js
@@ -22,24 +22,39 @@ exports.addRoutes = function (app) {
 
     //add a task
     app.post("/api/task", function (request, response) {
-        //console.log(request.body);
-        request.body.dateAdded = new Date();
-        db.insert(request.body);
-        response.setHeader('ContentType', 'application/json');
-        response.send({});
+        if (request.session.authenticated) {
+            //console.log(request.body);
+            request.body.dateAdded = new Date();
+            db.insert(request.body);
+            response.setHeader('ContentType', 'application/json');
+            response.send({});
+        } else {
+            response.status(401);
+            response.send();
+        }
     });
 
     //edits a task
     app.post("/api/task/:id", function (request, response) {
-        response.setHeader('ContentType', 'application/json');
-        response.send({});
+        if (request.session.authenticated) {
+            response.setHeader('ContentType', 'application/json');
+            response.send({});
+        } else {
+            response.status(401);
+            response.send();
+        }
     });
 
     //delete a task
     app.delete("/api/task/:id", function (request, response) {
-        console.log('id = ' + request.params.id);
-        db.remove({_id : request.params.id});
-        response.setHeader('ContentType', 'application/json');
-        response.send();
+        if (request.session.authenticated) {
+            console.log('id = ' + request.params.id);
+            db.remove({_id : request.params.id});
+            response.setHeader('ContentType', 'application/json');
+            response.send();
+        } else {
+            response.status(401);
+            response.send();
+        }
     });    
 }
